Name offset is now int, and removed boxed accessors of x,y coords of label and PNO

+++ src/dk/aau/cs/gui/undo/UpdateNameLabelOffsetCommand.java
@@ -4,10 +4,10 @@ import pipe.gui.graphicElements.PetriNetObject;
 import pipe.gui.graphicElements.PetriNetObjectWithLabel;
 
 public class UpdateNameLabelOffsetCommand extends Command {
-	double newXOffset, newYOffset, oldXOffset, oldYOffset;
+	int newXOffset, newYOffset, oldXOffset, oldYOffset;
 	PetriNetObjectWithLabel obj;
 	
-	public UpdateNameLabelOffsetCommand(double newXOffset, double newYOffset, double oldXOffset, double oldYOffset, PetriNetObjectWithLabel obj) {
+	public UpdateNameLabelOffsetCommand(int newXOffset, int newYOffset, int oldXOffset, int oldYOffset, PetriNetObjectWithLabel obj) {
 		this.obj = obj;
 		this.newXOffset = newXOffset;
 		this.newYOffset = newYOffset;
